requirements.txt
asgiref==3.7.2
channels==4.0.0
channels-redis==4.1.0
Django==4.2.5
djangorestframework==3.14.0
msgpack==1.0.5
Pillow==10.0.0
pytz==2023.3
redis==5.0.0
sqlparse==0.4.4
